Colorize Fragment Based On Price Change Percent
These changes will color the chart and ticker to red or
green depending on whether there were gains or losses.
To do this, we need to ensure that we get the response
from the ticker API request before rendering the chart. To
accomplish this, we enqueue the chartDataRequest inside
of the onResponse method for the ticker request.