Initial Redesign
These changes are an initial redesign of the app. There will now only be 2 tabs, one
for news and one for price+chart data. This will allow both the news and chart+price
user experiences to be richer since they wont both be crammed onto the screen at the
same time.