Fix Edge Case
If we remove an item from the favorites list by swiping it away, then we immediately attempt to re-add it from the AddFavoriteCoinActivity, returning to the favorites list after re-adding the coin would not automatically update the favorites list. This is because hashtables were being used in the smart comparisons which were not being updated by the adapter. These changes fix that.