Keep SearchView State On Tab Change
These changes will ensure that the state of the SearchView in the AllCurrencyListFragment is saved when save the user's search query in the AllCurrencyListFragment and restore it when they navigate back to the tab. If the user was in the SearchView before changing tabs, then when they navigate back, their previous search query will be restored, the keyboard will pop up and the searchview will gain focus!