Fix Nasty Bug
The onFavoriteChangeListener was getting called while SCROLLING in the recycler view because I am setting the favorite status in the view holder. I expected onFavoriteChangeListener to work just like onClickListener, where it would not be called unless a user physically pressed the button. Obviously that was not the case and the author of MaterialFavoriteButton does not make that clear.

Now I have changed the onFavoriteChangedListener to an onClickListener that sets the favorite manually. This ensures that the animation still occurs, but only when the user physically presses the button. What a doozy.