Fix Chart URL Swapping
Change the way the Chart URL is handled. Rather than formatting the URL at the service layer, I am formatting it in the fragment right when the user hits one of the interval buttons. Then I store the URL in a static var which is referenced by the service. This cleans up the code in the service and fixes the functionality. Also changed the default interval to 24h